Add legacy log uploader script
We iterate over all logs directly from Postgres, collating a new batch
of serialised log files every time the Hour part of the timestamp
changes.

The generated keyname is a hash of the data as a rudimentary attempt to
deduplicate logs. If the serialisation format changes, it's worth
deleting all old logs from S3 and re-uploading.